[jsk_recognition_utils] Separate cylinder.h from geo_util.h
