+++ jsk_recognition_utils/src/geo_util.cpp
@@ -408,74 +408,4 @@ namespace jsk_recognition_utils
     return ret;
   }
 
-  Cylinder::Cylinder(Eigen::Vector3f point, Eigen::Vector3f direction, double radius):
-    point_(point), direction_(direction), radius_(radius)
-  {
-
-  }
-
-  void Cylinder::filterPointCloud(const pcl::PointCloud<pcl::PointXYZ>& cloud,
-                                  const double threshold,
-                                  pcl::PointIndices& output)
-  {
-    Line line(direction_, point_);
-    output.indices.clear();
-    for (size_t i = 0; i < cloud.points.size(); i++) {
-      Eigen::Vector3f p = cloud.points[i].getVector3fMap();
-      double d = line.distanceToPoint(p);
-      if (d < radius_ + threshold && d > radius_ - threshold) {
-        output.indices.push_back(i);
-      }
-    }
-  }
-
-  void Cylinder::estimateCenterAndHeight(const pcl::PointCloud<pcl::PointXYZ>& cloud,
-                                         const pcl::PointIndices& indices,
-                                         Eigen::Vector3f& center,
-                                         double& height)
-  {
-    Line line(direction_, point_);
-    Vertices points;
-    for (size_t i = 0; i < indices.indices.size(); i++) {
-      int point_index = indices.indices[i];
-      points.push_back(cloud.points[point_index].getVector3fMap());
-    }
-    PointPair min_max = line.findEndPoints(points);
-    Eigen::Vector3f min_point = min_max.get<0>();
-    Eigen::Vector3f max_point = min_max.get<1>();
-    Eigen::Vector3f min_point_projected, max_point_projected;
-    line.foot(min_point, min_point_projected);
-    line.foot(max_point, max_point_projected);
-    height = (min_point_projected - max_point_projected).norm();
-    center = (min_point_projected + max_point_projected) / 2.0;
-  }
-
-  void Cylinder::toMarker(visualization_msgs::Marker& marker,
-                          const Eigen::Vector3f& center,
-                          const Eigen::Vector3f& uz,
-                          const double height)
-  {
-    marker.type = visualization_msgs::Marker::CYLINDER;
-    marker.pose.position.x = center[0];
-    marker.pose.position.y = center[1];
-    marker.pose.position.z = center[2];
-    Eigen::Vector3f orig_z(0, 0, 1);
-    Eigen::Quaternionf q;
-    q.setFromTwoVectors(orig_z, uz);
-    marker.pose.orientation.x = q.x();
-    marker.pose.orientation.y = q.y();
-    marker.pose.orientation.z = q.z();
-    marker.pose.orientation.w = q.w();
-    marker.scale.x = radius_ * 2;
-    marker.scale.y = radius_ * 2;
-    marker.scale.z = height;
-    marker.color.a = 1.0;
-    marker.color.g = 1.0;
-    marker.color.b = 1.0;
-  }
-
-  Eigen::Vector3f Cylinder::getDirection()
-  {
-    return direction_;
-  }
 }
